minify.pl improved to handle line breaks in non-HTML code, e.g. CSS, by replacing each sequence of one or more line breaks with one space. (Perl's HTML::TreeBuilder->as_HTML("","",\%empty) minifies HTML code but not CSS, etc., which it does not parse.)  Sequences of spaces and tabs not at the beginning of a line are not replaced with single spaces in unparsed, non-HTML code, so that code is probably not fully minified. Also, it will break a here-document spanning multiple lines in non-HTML code.
subscriberToPublic.xsl employs article title (linked text) prefixes in an XHTML version of the subscriber's new issue index page to determine the link destinations for the output public index page for the new issue.  Instructions are in subscriberToPublicVars.xsl, which now has only one variable which must be edited for each new issue.